+++ data/user_agreement.txt
@@ -1,0 +1,25 @@
+By using this application, you agree to the following terms and conditions:
+
+1. You must be at least 18 years old to use this application.
+2. You agree to abide by all applicable laws and regulations.
+3. Your personal information may be collected and used as described in our privacy policy.
+4. We are not responsible for any damages or losses incurred while using this application.
+5. You are solely responsible for the content you upload or share through this application.
+
+Please read this agreement carefully before using the application.
+If you do not agree with any of these terms, please do not use the application.
+--------------------------------------------------------------------------------------------------------
+
+Privacy Policy
+
+We collect information about you when you use this application.
+This information is used to provide a better user experience and improve our services.
+We may share your personal information with third parties for marketing purposes.
+You can opt out of these communications at any time by contacting
+us at contact@samplecompany.com
+--------------------------------------------------------------------------------------------------------
+
+Contact Us
+
+If you have any questions about this agreement, please contact
+us at contact@samplecompany.com.
