+++ weather_index.json
@@ -1,4 +1,3 @@
-
 {
   "clear_sky": 0,
   "cloudy": [1, 2, 3],
